backend/requirements.txt
fastapi
uvicorn
openai
finnhub-python
python-dotenv
pydantic
langchain
langchain-core
langchain-community
langgraph
chromadb
sentence-transformers
pypdf
bcrypt
python-jose[cryptography]
sqlalchemy
python-multipart